requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
pypdf
tiktoken
jinja2